Split defines into a .h
